apis: move code around to avoid a circular import...
... that only happens in production, because REASONS.
